Refactor attributes into attribute package
