+++ src/test/java/seedu/address/logic/commands/ViewDayCommandTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.task.Date;
+import seedu.address.model.task.attributes.Date;
 import seedu.address.model.task.predicates.TaskOnDatePredicate;
 
 class ViewDayCommandTest {
