+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -36,11 +36,11 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.model.tag.Tag;
-import seedu.address.model.task.Date;
-import seedu.address.model.task.Duration;
-import seedu.address.model.task.RecurringSchedule;
+import seedu.address.model.task.attributes.Date;
+import seedu.address.model.task.attributes.Duration;
+import seedu.address.model.task.attributes.RecurringSchedule;
 import seedu.address.model.task.Task;
-import seedu.address.model.task.Title;
+import seedu.address.model.task.attributes.Title;
 import seedu.address.testutil.TaskBuilder;
 
 public class AddCommandParserTest {
