+++ src/test/java/seedu/address/model/task/DateTest.java
@@ -6,6 +6,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.task.attributes.Date;
+
 public class DateTest {
 
     @Test
