+++ src/test/java/seedu/address/model/task/DescriptionTest.java
@@ -5,6 +5,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.task.attributes.Description;
+
 public class DescriptionTest {
 
     @Test
