+++ src/test/java/seedu/address/model/task/StatusTest.java
@@ -8,6 +8,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.task.attributes.Status;
+
 public class StatusTest {
 
     @Test
