+++ src/test/java/seedu/address/model/task/TitleTest.java
@@ -6,6 +6,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.task.attributes.Title;
+
 public class TitleTest {
 
     @Test
