+++ src/test/java/seedu/address/model/task/predicates/TaskDateOnDatePredicateTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.model.task.Date;
+import seedu.address.model.task.attributes.Date;
 import seedu.address.testutil.TaskBuilder;
 
 class TaskDateOnDatePredicateTest {
